+++ src/FolkerKinzel.VCards.Tests/V3Tests.cs
@@ -12,9 +12,6 @@ public class V3Tests
     [TestMethod]
     public void ParseTest()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         IList<VCard>? vcard = Vcf.Load(TestFiles.V3vcf);
 
         Assert.IsNotNull(vcard);
@@ -25,9 +22,6 @@ public class V3Tests
     [TestMethod]
     public void ParseTest2()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         IList<VCard>? vcard = Vcf.Load(@"C:\Users\fkinz\OneDrive\Kontakte\Thunderbird\21-01-13.vcf");
 
         Assert.IsNotNull(vcard);
@@ -37,9 +31,6 @@ public class V3Tests
     [TestMethod]
     public void ParseTest3()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         IList<VCard>? vcard = Vcf.Load(TestFiles.PhotoV3vcf);
 
         Assert.IsNotNull(vcard);
@@ -50,9 +41,6 @@ public class V3Tests
     [TestMethod]
     public void WriteEmptyVCardTest()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         var vcard = new VCard();
 
         string s = vcard.ToVcfString(VCdVersion.V3_0);
@@ -77,9 +65,6 @@ public class V3Tests
     [TestMethod]
     public void LineWrappingTest()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         var vcard = new VCard();
 
         string UNITEXT = "Dies ist ein wirklich sehr sehr sehr langer Text mit ü, Ö, und ä " + "" +
@@ -136,9 +121,6 @@ public class V3Tests
     [TestMethod]
     public void SaveTest()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         var vcard = new VCard();
 
         string UNITEXT = "Dies ist ein wirklich sehr sehr sehr langer Text mit ü, Ö, und ä " + "" +
@@ -227,9 +209,6 @@ public class V3Tests
     [TestMethod]
     public void TimeDataTypeTest()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         const string vcardString = @"BEGIN:VCARD
 VERSION:3.0
 BDAY;Value=Time:05:30:00
@@ -255,9 +234,6 @@ END:VCARD";
     [TestMethod]
     public void NonStandardParameterTest1()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         const string whatsAppNumber = "+1-234-567-89";
         var xiamoiMobilePhone = new TextProperty(whatsAppNumber);
         xiamoiMobilePhone.Parameters.NonStandard =
@@ -286,9 +262,6 @@ END:VCARD";
     [TestMethod]
     public void ImppTest1()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         const string mobilePhoneNumber = "tel:+1-234-567-89";
         var whatsAppImpp = new TextProperty(mobilePhoneNumber);
 
@@ -316,9 +289,6 @@ END:VCARD";
     [TestMethod]
     public void ImppTest2()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         const string mobilePhoneNumber = "tel:+1-234-567-89";
         var whatsAppImpp = new TextProperty(mobilePhoneNumber);
 
@@ -344,11 +314,22 @@ END:VCARD";
     }
 
     [TestMethod]
+    public void ImppTest3() 
+    {
+        const string serialized = """
+            BEGIN:VCARD
+            VERSION:3.0
+            IMPP;TYPE=OTHER:abc
+            END:VCARD
+            """;
+
+        VCard vc = Vcf.Parse(serialized)[0];
+        Assert.IsNotNull(vc.Messengers?.First()!.Parameters.NonStandard);
+    }
+
+    [TestMethod]
     public void XMessengerTest1()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         const string mobilePhoneNumber = "skype:+1-234-567-89";
         var prop = new TextProperty(mobilePhoneNumber);
         const Impp imppTypes = Impp.Mobile | Impp.Personal;
@@ -374,9 +355,6 @@ END:VCARD";
     [TestMethod]
     public void XMessengerTest2()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         const string mobilePhoneNumber = "skype:+1-234-567-89";
         var prop = new TextProperty(mobilePhoneNumber);
         const Impp imppTypes = Impp.Mobile | Impp.Personal;
@@ -398,9 +376,6 @@ END:VCARD";
     [TestMethod]
     public void PreserveTimeZoneAndGeoTest1()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         var adr = new AddressProperty("1", "", "", "", "");
         adr.Parameters.TimeZone = TimeZoneID.Parse("Europe/Berlin");
         adr.Parameters.GeoPosition = new GeoCoordinate(52, 13);
@@ -422,9 +397,6 @@ END:VCARD";
     [TestMethod]
     public void PreserveTimeZoneAndGeoTest2()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         var adr = new AddressProperty("1", "", "", "", "");
         adr.Parameters.TimeZone = TimeZoneID.Parse("Europe/Berlin");
         adr.Parameters.GeoPosition = new GeoCoordinate(52, 13);
@@ -447,9 +419,6 @@ END:VCARD";
     [TestMethod]
     public void DisplayNameTest1()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         var nm = new NameProperty("Kinzel", "Folker");
 
         var vCard = new VCard { NameViews = nm };
@@ -466,9 +435,6 @@ END:VCARD";
     [TestMethod]
     public void DisplayNameTest2()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         var vCard = new VCard { NameViews = [null] };
 
         string vcf = vCard.ToVcfString(VCdVersion.V3_0);
@@ -483,9 +449,6 @@ END:VCARD";
     [TestMethod]
     public void DisplayNameTest3()
     {
-        VCard.SyncTestReset();
-        VCard.RegisterApp(null);
-
         var vCard = new VCard();
 
         string vcf = vCard.ToVcfString(VCdVersion.V3_0, options: Opts.Default.Set(Opts.WriteEmptyProperties));
@@ -679,4 +642,6 @@ END:VCARD";
         Assert.AreEqual(Adr.Dom | Adr.Postal | Adr.Parcel,
                         vc.Addresses!.First()!.Parameters.AddressType);
     }
+
+    
 }